Building with CMake [WIP]
Turns out the only C++ IDE for Linux that I like (CLion) only works right for CMake projects.

The ideal solution would be implementing a premake5 module to generate a CMake but I don't have the stomach for digging into that undocumented mess for now, so this ad-hoc term solution should do.

TODO (optional): make it place the generated libraries and tests in the right directories