Prototype client class
There are some things wrong with the current approach. The worst of them is that the negotiation never completes as intended: the server gets a valid header first, enters into kConnection state and forgets about sending its part of the negotiation.

A new state kChallenge is needed. I'll give it a few thoughts.